lint errors and move ensureWebhook func to cc controller
